appj_requirements.txt
matplotlib==3.5.3
numpy
scipy
ipykernel
seabreeze
libusb1
pyserial
opencv-python==4.7.0.72
pyvisa
python-usbtmc
pyusb
crcmod
